Cleanup core, move test_utils to core
