--- aria/core/viewsets.py
@@ -1,31 +1,0 @@
-from rest_framework import status
-from rest_framework.permissions import AllowAny
-from rest_framework.response import Response
-from rest_framework.views import APIView
-
-from rest_framework_simplejwt.tokens import RefreshToken
-from rest_framework_simplejwt.views import TokenObtainPairView
-
-from aria.core.serializers import MyTokenObtainPairSerializer
-
-
-class MyTokenObtainPairView(TokenObtainPairView):
-    serializer_class = MyTokenObtainPairSerializer
-
-
-class LogoutAndBlacklistRefreshTokenForUserView(APIView):
-    """
-    Viewset for blacklisting used and expired refresh tokens
-    """
-
-    permission_classes = (AllowAny,)
-    authentication_classes = ()
-
-    def post(self, request):
-        try:
-            refresh_token = request.data["refresh_token"]
-            token = RefreshToken(refresh_token)
-            token.blacklist()
-            return Response(status=status.HTTP_205_RESET_CONTENT)
-        except RuntimeError:
-            return Response(status=status.HTTP_400_BAD_REQUEST)
