+++ aria/users/tests/test_internal_serializers.py
@@ -1,7 +1,7 @@
 import pytest
 from model_bakery import baker
 
-from aria.test_utils import model_baker_datetime_formatting
+from aria.core.test_utils import model_baker_datetime_formatting
 from aria.users.models import User
 from aria.users.viewsets.internal import UserDetailAPI, UserListAPI, UserUpdateAPI
